--- dpr1/src/test/ConexionServidor.java
@@ -1,47 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package test;
-
-import java.io.IOException;
-import java.io.ObjectOutputStream;
-import java.net.Socket;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
-/**
- *
- * @author david
- */
-public class ConexionServidor {
-    private Socket socket;
-    private ObjectOutputStream oos;
-
-    public Socket getSocket() {
-        return socket;
-    }
-
-    public void setSocket(Socket socket) {
-        this.socket = socket;
-    }
-
-    public ObjectOutputStream getOos() {
-        return oos;
-    }
-
-    public void setOos(ObjectOutputStream oos) {
-        this.oos = oos;
-    }
-    
-    public ConexionServidor(Socket socket){
-        this.socket = socket;
-        try {
-            oos = new ObjectOutputStream(socket.getOutputStream());
-        } catch (IOException ex) {
-            Logger.getLogger(ConexionServidor.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-}
